+++ src/main/java/net/thesquire/backroomsmod/block/custom/ModDoorBlock.java
@@ -4,8 +4,10 @@ import net.minecraft.block.Block;
 import net.minecraft.block.BlockSetType;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.DoorBlock;
+import net.minecraft.block.enums.DoubleBlockHalf;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.ItemPlacementContext;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.state.StateManager;
@@ -52,6 +54,16 @@ public class ModDoorBlock extends DoorBlock {
         this.playOpenCloseSound(player, world, pos, state.get(OPEN));
         world.emitGameEvent(player, this.isOpen(state) ? GameEvent.BLOCK_OPEN : GameEvent.BLOCK_CLOSE, pos);
         return ActionResult.success(world.isClient);
+    }
+
+    @Override
+    public void neighborUpdate(BlockState state, World world, BlockPos pos, Block sourceBlock, BlockPos sourcePos, boolean notify) {
+        if (state.get(LOCKED) && !state.get(OPEN)) {
+            PlayerEntity player = world.getClosestPlayer(sourcePos.getX(), sourcePos.getY(), sourcePos.getZ(), 16, false);
+            if (player != null) player.sendMessage(Text.translatable("door.locked"), true);
+            return;
+        }
+        super.neighborUpdate(state, world, pos, sourceBlock, sourcePos, notify);
     }
 
     private void playOpenCloseSound(@Nullable Entity entity, World world, BlockPos pos, boolean open) {