+++ src/main/java/net/thesquire/backroomsmod/block/ModBlocks.java
@@ -4,6 +4,8 @@ import net.fabricmc.fabric.api.item.v1.FabricItemSettings;
 import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
 import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
 import net.minecraft.block.*;
+import net.minecraft.block.enums.Instrument;
+import net.minecraft.block.piston.PistonBehavior;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.item.TooltipContext;
 import net.minecraft.item.BlockItem;
@@ -29,21 +31,36 @@ import java.util.List;
 
 public class ModBlocks {
 
+    private static final FabricBlockSettings STONE = FabricBlockSettings.create()
+            .mapColor(MapColor.STONE_GRAY)
+            .instrument(Instrument.BASEDRUM);
+    private static final FabricBlockSettings METAL = FabricBlockSettings.create()
+            .mapColor(MapColor.IRON_GRAY)
+            .instrument(Instrument.IRON_XYLOPHONE)
+            .sounds(BlockSoundGroup.METAL);
+    private static final FabricBlockSettings GLOWSTONE = FabricBlockSettings.create()
+            .mapColor(MapColor.PALE_YELLOW)
+            .instrument(Instrument.PLING)
+            .strength(0.3F)
+            .sounds(BlockSoundGroup.GLASS)
+            .luminance((state) -> 15)
+            .solidBlock(Blocks::never);
+
     public static final Block BISMUTHINITE_ORE = registerBlock("bismuthinite_ore",
-            new Block(FabricBlockSettings.of(Material.STONE).strength(4.5f).requiresTool()));
+            new Block(STONE.strength(4.5f).requiresTool()));
 
     public static final Block CEILING_TILE = registerBlock("ceiling_tile",
-            new CeilingTileBlock(FabricBlockSettings.of(Material.STONE).strength(1.6f).requiresTool()));
+            new CeilingTileBlock(STONE.strength(1.6f).requiresTool()));
 
     public static final Block YELLOW_WALLPAPER = registerBlock("yellow_wallpaper",
-            new Block(FabricBlockSettings.of(Material.STONE).strength(1.6f).requiresTool()));
+            new Block(STONE.strength(1.6f).requiresTool()));
 
     public static final Block TFMC_MAGNET = registerBlock("tfmc_magnet",
-            new TFMCMagnetBlock(FabricBlockSettings.of(Material.METAL).strength(4f).requiresTool()),
+            new TFMCMagnetBlock(METAL.strength(4f).requiresTool()),
             "block.backroomsmod.tfmc_magnet.tooltip_1", "block.backroomsmod.tfmc_magnet.tooltip_2");
 
     public static final Block WAREHOUSE_CONCRETE = registerBlock("warehouse_concrete",
-            new WarehouseConcreteBlock(FabricBlockSettings.of(Material.STONE).strength(1.8f).requiresTool()));
+            new WarehouseConcreteBlock(STONE.strength(1.8f).requiresTool()));
 
     public static final Block PAINTED_WAREHOUSE_CONCRETE = registerBlock("painted_warehouse_concrete",
             new PaintedWarehouseConcreteBlock(FabricBlockSettings.copy(WAREHOUSE_CONCRETE)));
@@ -58,7 +75,7 @@ public class ModBlocks {
             new PipeBlock(FabricBlockSettings.copy(Blocks.IRON_BLOCK)));
 
     public static final Block IRON_BUTTON = registerBlock("iron_button",
-            new ButtonBlock(FabricBlockSettings.of(Material.DECORATION).noCollision().strength(0.5f).requiresTool(), BlockSetType.IRON, 20, false));
+            new ButtonBlock(FabricBlockSettings.create().noCollision().strength(0.5f).requiresTool().pistonBehavior(PistonBehavior.DESTROY), BlockSetType.IRON, 20, false));
 
     // Blocks with a GUI or BlockEntity have to be registered in the method below to ensure proper register order!
     public static Block INDUSTRIAL_ALLOY_SMELTER;
@@ -79,17 +96,14 @@ public class ModBlocks {
                 new GenericMachineBlock(ModGuis.MAGNETIC_DISTORTION_SYSTEM_CONTROL_COMPUTER, MagneticDistortionSystemControlComputerBlockEntity::new));
 
         FLUORESCENT_LIGHT = registerBlock("fluorescent_light",
-                new FluorescentLightBlock(FabricBlockSettings.of(Material.GLASS)
-                        .strength(0.3f).requiresTool().sounds(BlockSoundGroup.GLASS)
-                        .luminance(FluorescentLightBlock::getLuminance)));
+                new FluorescentLightBlock(GLOWSTONE.requiresTool().luminance(FluorescentLightBlock::getLuminance)));
 
         MOUNTABLE_FLUORESCENT_LIGHT = registerBlock("mountable_fluorescent_light",
-                new MountableFluorescentLightBlock(FabricBlockSettings.of(Material.GLASS).strength(0.3f).requiresTool()
-                        .nonOpaque().sounds(BlockSoundGroup.GLASS).solidBlock(ModBlocks::never).suffocates(ModBlocks::never)
+                new MountableFluorescentLightBlock(GLOWSTONE.requiresTool().nonOpaque().suffocates(ModBlocks::never)
                         .blockVision(ModBlocks::never).luminance(MountableFluorescentLightBlock::getLuminance)), 16);
 
         PORTAL_PLACER = registerBlockWithoutBlockItem("portal_placer",
-                new PortalPlacerBlock(FabricBlockSettings.of(Material.AIR).noCollision().dropsNothing().noBlockBreakParticles()));
+                new PortalPlacerBlock(FabricBlockSettings.create().noCollision().dropsNothing().noBlockBreakParticles()));
 
         ELEVATOR_BUTTON = registerBlockWithoutBlockItem("elevator_button",
                 new ElevatorButton(FabricBlockSettings.copy(IRON_BUTTON), BlockSetType.IRON, 20, false));
@@ -118,13 +132,13 @@ public class ModBlocks {
     private static void registerBlockItem(String name, Block block) {
         Item item = Registry.register(Registries.ITEM, BackroomsMod.makeId(name),
                 new BlockItem(block, new FabricItemSettings()));
-        ItemGroupEvents.modifyEntriesEvent(ModItemGroup.BACKROOMS).register(entries -> entries.add(item));
+        ItemGroupEvents.modifyEntriesEvent(ModItemGroup.BACKROOMS_ITEM_GROUP).register(entries -> entries.add(item));
     }
 
     private static void registerBlockItem(String name, Block block, int stackSize) {
         Item item = Registry.register(Registries.ITEM, BackroomsMod.makeId(name),
                 new BlockItem(block, new FabricItemSettings().maxCount(stackSize)));
-        ItemGroupEvents.modifyEntriesEvent(ModItemGroup.BACKROOMS).register(entries -> entries.add(item));
+        ItemGroupEvents.modifyEntriesEvent(ModItemGroup.BACKROOMS_ITEM_GROUP).register(entries -> entries.add(item));
     }
 
     private static void registerBlockItem(String name, Block block, String... tooltipKeys) {
@@ -141,7 +155,7 @@ public class ModBlocks {
                         }
                     }
                 });
-        ItemGroupEvents.modifyEntriesEvent(ModItemGroup.BACKROOMS).register(entries -> entries.add(item));
+        ItemGroupEvents.modifyEntriesEvent(ModItemGroup.BACKROOMS_ITEM_GROUP).register(entries -> entries.add(item));
     }
 
     private static Block registerBlockWithoutBlockItem(String name, Block block) {